Update directory depth check during init.
Working with sandbox base stack which is structured a little
differently.
Stack directory only needs to be 2 deep, not 3, as we use the first
directory and the last two, so the first and second last elements can
be the same directory.
Can now initialise stacks that are only 2 directories deep below the
nominated stack prefix, instead of 3.